rag/requirements.txt
# Core LangChain (User-specified stable versions for Python 3.11)
langchain==0.3.26
langchain-pinecone==0.2.8
langchain-openai==0.3.24
langchain-community==0.3.26

# Embeddings (Fast sentence-transformers)
sentence-transformers==4.1.0
# torch==2.5.1
transformers==4.47.1

# Vector Database
pinecone>=7.0.0,<8.0.0

# PDF Processing (Multiple engines for speed and reliability)
pypdf==5.6.1
PyMuPDF==1.24.0  # Faster alternative to pypdf

# OCR Dependencies
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==10.4.0

# LLM Client (Ollama)
requests==2.32.3

# Text Processing
tiktoken==0.8.0

# Utils
python-dotenv==1.1.0
numpy==1.26.4

# Performance
orjson==3.10.12  # Fast JSON parsing

# Install in editable mode
-e .
